init tcp module
