+++ huginn-net-tcp/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "huginn-net-tcp"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.80"
+description = "TCP fingerprinting and analysis for Huginn Net"
+license = "MIT OR Apache-2.0"
+authors = ["Maximiliano Biandratti"]
+repository = "https://github.com/biandratti/huginn-net"
+keywords = ["tcp", "fingerprinting", "network", "security"]
+categories = ["network-programming"]
+publish = true
+
+[dependencies]
+huginn-net-db = { path = "../huginn-net-db" }
+tracing = { workspace = true }
+thiserror = { workspace = true }
+pnet = { workspace = true }
+ttl_cache = { workspace = true }
